% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmd_scaled_rmttest_function.R
\name{fmd_scaled_rmttest}
\alias{fmd_scaled_rmttest}
\title{Allometric scaling for flow-mediated dilation: Repeated measures t-tests}
\usage{
fmd_scaled_rmttest(dat)
}
\arguments{
\item{dat}{data frame object; does not have to be named 'dat', but the columns that correspond to peak artery diameter, baseline artery diameter, and group (or condition) must be labelled "dpeak", "dbase", and "group", respectively.}
}
\value{
This function returns the following:
\item{model.coef}{A dataframe continaing coefficients from the linear model}
\item{main.effects}{A dataframe containing main effects and statistical contrasts}
\item{transformed.emmeans}{Backtransformed estimated marginal means and model standard error}
\item{plot}{Graphical representation using \code{ggplot2} syntax}
}
\description{
This function calculates allometrically-scaled flow-mediated dilation responses for repeated measures t-test study designs, and returns  comparisons between groups. The function will also return back-transformed means, standard errors, and 95\% confidence intervals. You need to label the columns in your data set as "dpeak", "dbase", and "group" respectively.
}
\examples{
Create a sample data frame (simulated data; as such, values may not correspond with physiological norms) with peak and baseline artery diameters from sample participants:
dat <- data.frame(pid = as.factor(rep(1:10, 2)), dpeak = rnorm(20, 4.27, 1.12), dbase = rnorm(20, 4.16, 1.21)*0.8, group = as.factor(c(rep("NS", 10), rep("SD", 10))))
dat <- dat |>
dplyr::arrange(pid)
res <- fmd_scaled_rmttest(dat)

}
\keyword{FMD;}
\keyword{allometric}
\keyword{scaled;}
